added registration functionality
