+++ routes/user.js
@@ -1,8 +1,9 @@
 const express = require("express");
-const { login } = require("../controllers/user");
+const { login, register } = require("../controllers/user");
 
 const router = express.Router();
 
 router.post("/login", login);
+router.post("/register", register);
 
 module.exports = router;
